pyproject.toml
[build-system]
# editable install support (PEP 660) added in 64.0
requires = ["setuptools >= 64.0.0"]
build-backend = "setuptools.build_meta"

[project]
version = "0.0.1"
name = "writertools"
description = "A collection of tools for writers"
readme = "README.md"
requires-python = ">=3.8"
# keywords = ["one", "two"]
license = { text = "Apache 2.0" }
classifiers = ["Framework :: Django", "Programming Language :: Python :: 3"]
dependencies = [
    "django >= 3.2.17, < 4.0",
    "django-allauth >= 0.52.0",
    "django-bootstrap5 >= 22.2",
    "django-bootstrap-icons >= 0.8.2",
    "django-environ >= 0.9.0",
    "docutils >= 0.19",
    "pillow >= 9.4.0",
    "django-genericsite @ https://github.com/veselosky/django-genericsite/releases/download/release-0.1.3/django_genericsite-0.1.3-py3-none-any.whl",
]

[project.optional-dependencies]
dev = [
    "black",
    "django_debug_toolbar",
    "django_extensions",
    "pytest >= 6.2.5",
    "pytest-django >= 4.5.2",
]

# [project.scripts]
# my-script = "my_package.module:function"

[tool.setuptools]
packages = ["writertools"]

[tool.pytest.ini_options]
minversion = "6.2.5"
DJANGO_SETTINGS_MODULE = "writertools.settings"
# addopts = "-ra -q"
# testpaths = ["tests", "integration"]
python_files = ["tests.py", "test_*.py", "*_test.py", "*_tests.py"]

# By default pytest will show all warnings including DeprecationWarning.
# https://docs.python.org/3/library/warnings.html#the-warnings-filter
# TLDR: ignore:$message:$category:$module:$lineno
filterwarnings = [
    # removed in Django 4.1, but required for <3.2
    "ignore:.*default_app_config::",
]
